Pull terrain images into the project and update previews
